Added already trained embeddings
